+++ wasmbinding/test/testdata/cw_nameservice.wasm
